Rename previewState -> preview
