+++ src/store/modules/index.js
@@ -7,18 +7,18 @@ import type { Location } from 'react-router'
 import entities from './entities'
 import exportAsPdf from './exportAsPdf'
 import editorState from './editorState'
-import previewState from './previewState'
+import preview from './preview'
 
 import type { EntitiesState } from './entities'
 import type { ExportAsPdfState } from './exportAsPdf'
 import type { EditorState } from './editorState'
-import type { PreviewState } from './previewState'
+import type { PreviewState } from './preview'
 
 export type RootState = {
   entities: EntitiesState,
   exportAsPdf: ExportAsPdfState,
   editorState: EditorState,
-  previewState: PreviewState,
+  preview: PreviewState,
   router: ?Location,
 }
 
@@ -26,6 +26,6 @@ export default combineReducers({
   entities,
   exportAsPdf,
   editorState,
-  previewState,
+  preview,
   router: routerReducer,
 })
